+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.7.0.0
 
+* Add instances for `InjectRuleFailure` and switch to using `injectFailure`
+* Add `allegraToAlonzoUtxoPredFailure`
 * Add `NFData` instances for:
   * `CollectError`
   * `AlonzoUtxoPredFailure`